cli documentation; move low level fns
